Changed notification file extension.
